+++ src/asphalt/core/__init__.py
@@ -25,7 +25,6 @@ from ._event import Signal as Signal
 from ._event import SignalQueueFull as SignalQueueFull
 from ._event import stream_events as stream_events
 from ._event import wait_event as wait_event
-from ._exceptions import ApplicationExit as ApplicationExit
 from ._exceptions import AsyncResourceError as AsyncResourceError
 from ._exceptions import NoCurrentContext as NoCurrentContext
 from ._exceptions import ResourceConflict as ResourceConflict